third_party/WebKit/LayoutTests/inspector/console/console-uncaught-promise-expected.txt

CONSOLE ERROR: line 21: Uncaught (in promise) Error: err1
CONSOLE ERROR: line 40: Uncaught (in promise) Error: err2
CONSOLE ERROR: line 56: Uncaught (in promise) NotFoundError: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
CONSOLE ERROR: line 61: Uncaught (in promise) 42
CONSOLE ERROR: line 65: Uncaught (in promise) 1e+100
CONSOLE ERROR: line 69: Uncaught (in promise) foo
CONSOLE ERROR: line 73: Uncaught (in promise) #<Object>
CONSOLE ERROR: line 77: Uncaught (in promise) undefined
CONSOLE ERROR: line 81: Uncaught (in promise) SecurityError: Failed to register a ServiceWorker: The URL protocol of the current origin ('') is not supported.
Tests that uncaught promise rejections are logged into console.

console-uncaught-promise.html:21 Uncaught (in promise) Error: err1
promiseTest1 @ console-uncaught-promise.html:21
timeout @ console-uncaught-promise.html:15
Promise rejected (async)
promiseTest1 @ console-uncaught-promise.html:21
timeout @ console-uncaught-promise.html:15
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:14
(anonymous) @ VM:1
console-uncaught-promise.html:40 Uncaught (in promise) Error: err2
promiseTest2 @ console-uncaught-promise.html:40
timeout @ console-uncaught-promise.html:15
Promise rejected (async)
(anonymous) @ console-uncaught-promise.html:36
Promise rejected (async)
(anonymous) @ console-uncaught-promise.html:35
Promise rejected (async)
(anonymous) @ console-uncaught-promise.html:34
Promise rejected (async)
promiseTest2 @ console-uncaught-promise.html:33
timeout @ console-uncaught-promise.html:15
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:14
(anonymous) @ VM:1
console-uncaught-promise.html:56 Uncaught (in promise) DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
throwDOMException @ console-uncaught-promise.html:56
catcher @ console-uncaught-promise.html:49
Promise rejected (async)
promiseTest3 @ console-uncaught-promise.html:48
timeout @ console-uncaught-promise.html:15
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:14
(anonymous) @ VM:1
console-uncaught-promise.html:61 Uncaught (in promise) 42
promiseTest4 @ console-uncaught-promise.html:61
timeout @ console-uncaught-promise.html:15
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:14
(anonymous) @ VM:1
console-uncaught-promise.html:65 Uncaught (in promise) 1e+100
promiseTest5 @ console-uncaught-promise.html:65
timeout @ console-uncaught-promise.html:15
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:14
(anonymous) @ VM:1
console-uncaught-promise.html:69 Uncaught (in promise) foo
promiseTest6 @ console-uncaught-promise.html:69
timeout @ console-uncaught-promise.html:15
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:14
(anonymous) @ VM:1
console-uncaught-promise.html:73 Uncaught (in promise) {foo: 42}
    foo: 42
    __proto__: Object
promiseTest7 @ console-uncaught-promise.html:73
timeout @ console-uncaught-promise.html:15
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:14
(anonymous) @ VM:1
console-uncaught-promise.html:77 Uncaught (in promise) undefined
promiseTest8 @ console-uncaught-promise.html:77
timeout @ console-uncaught-promise.html:15
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:14
(anonymous) @ VM:1
console-uncaught-promise.html:81 Uncaught (in promise) DOMException: Failed to register a ServiceWorker: The URL protocol of the current origin ('file://') is not supported.
promiseTest9 @ console-uncaught-promise.html:81
timeout @ console-uncaught-promise.html:15
Promise (async)
promiseTest9 @ console-uncaught-promise.html:81
timeout @ console-uncaught-promise.html:15
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:14
(anonymous) @ VM:1
